+++ neuronquant/network/transport.py
@@ -22,46 +22,75 @@ import zmq
 import time
 import json
 import datetime as dt
+import abc
 
 from neuronquant.utils.signals import SignalManager
-from neuronquant.utils import remote_setup, color_setup, setup
+from neuronquant.utils import remote_setup
 
 log = logbook.Logger('ZMQ Messaging')
-#https://github.com/JustinTulloss/zeromq.node/blob/master/examples/devices/queue.js
-#https://github.com/zeromq/pyzmq/blob/master/examples/device/server.py
 #TODO http://pythonhosted.org/Logbook/setups.html
 #     http://pythonhosted.org/Logbook/api/queues.html
-#TODO 'dashboard' destination hardcoded
 
 
-#TODO This class is abstract
+# Alias for acknowledgment messages
+OK_CODE = 0
+KO_CODE = 1
+
+
 class ZMQ_Base(object):
     '''
-    Abstract class common for every type of ZMQ device
+    Abstract class in common for every type of ZMQ device. Messages sent through
+    network between QuanTrade processes are json, and should look like:
+        {
+            'time': '14, March 2012 15:12',
+            'id': 'sender id',
+            'channel': 'receiver id',
+            'type': 'used by broker for filtering',
+            'msg': 'json or plain text informations'
+         }
+    Most of the time messages will ben sent to the broker which will read
+    those fields and route accordingly the message toward a client that
+    will process the informations.
     '''
-    def __init__(self, id=None, signal_manager=True):
+
+    __metaclass__ = abc.ABCMeta
+
+    def __init__(self, id=None, recipient='dashboard', signal_manager=True):
+        # Used in messages 'id' field, for authentification on receiver side
         self.identity = id
+
+        # So said receiver, allows the broker to route the message toward the right client
+        self.recipient = recipient
+
+        # Handles especially CTRL-C keyboard interruption or alarms (when waiting with a timeout)
         signals = [signal.SIGINT]
         self.signal_manager = SignalManager(signal_codes=signals)
         log.info(self.signal_manager)
+
+        # Every ZMQ sockets need to initialize a context
         self.context = zmq.Context()
-        self.port = None
-        self.socket = None
+
+        self.port    = None
+        self.socket  = None
 
     def receive(self, json=True, acknowledgment=None):
         msg = None
         msg = self.socket.recv_json() if json else self.socket.recv()
-        #msg = self.socket.recv()
         log.debug('ZMQ Agent received {}'.format(msg))
+
+        # If requested, we let the sender know the message was received
         if acknowledgment:
             if json:
+                # Complete protocole, default
                 self.socket.send_json({'time': dt.datetime.strftime(dt.datetime.now(), format='%Y-%m-%dT%H:%M:%S'),
-                                       'id': self.identity, 'channel': 'dashboard', 'msg': 0})
+                                       'id': self.identity, 'channel': self.recipient, 'msg': OK_CODE})
             else:
-                self.socket.send('{}:{}'.format(self.identity, 0))
+                # Simple string response
+                self.socket.send('{}:{}'.format(self.identity, OK_CODE))
         return msg
 
     def __del__(self):
+        #FIXME Correct seesion end
         #if not self.socket.closed:
         #self.socket.close()
         #self.context.term()
@@ -69,51 +98,86 @@ class ZMQ_Base(object):
 
 
 class ZMQ_Dealer(ZMQ_Base):
-    def run(self, uri=None, host='localhost', port=5555):
+    '''
+    ZMQ Dealer device, http://www.zeromq.org/tutorials:dealer-and-router
+    This is the client device, used by the simulator and the portfolio manager,
+    and meant (but not limited) to talk to the broker.
+    '''
+    #NOTE Default should be reade from ~/.quantrade/default.json (method from abstract class)
+    def run(self, uri=None, host='localhost', port=5570):
+        '''
+        Device setup and makes it listen
+        '''
+        # Create 'Dealer' type socket
         self.socket = self.context.socket(zmq.DEALER)
-        self.port = port
+        self.port   = port
+
+        # Set identity that will be communicated as authentification in messages
         self.socket.setsockopt(zmq.IDENTITY, self.identity if self.identity else str(self.port))
+
+        # If no explicit uri, we use given or default host and port
         if uri is None:
             uri = 'tcp://{}:{}'.format(host, port)
+
+        # Setup done, running
         self.socket.connect(uri)
+        # Non-blocking mechanism
         self.poll = zmq.Poller()
         self.poll.register(self.socket, zmq.POLLIN)
         log.info('Client connected to {}.'.format(uri))
 
     def noblock_recv(self, timeout=0, json=True, acknowledgment=None):
+        '''
+        Checks for pending messages on socket but won't wait for new to arrive
+        '''
+        # Checks
         socks = dict(self.poll.poll(timeout))
-        msg = None
+        msg   = None
+        # Something arrived for this device ?
         if self.socket in socks:
+            # A new message is pending ?
             if socks[self.socket] == zmq.POLLIN:
                 msg = self.socket.recv_json() if json else self.socket.recv()
                 log.debug('Client received {}'.format(msg))
+
                 if acknowledgment:
                     if json and msg:
-                        self.send('ok', type='acknowledgment')
+                        self.send(OK_CODE, type='acknowledgment')
                     elif json and not msg:
-                        self.send('ko', type='acknowledgment')
+                        self.send(KO_CODE, type='acknowledgment')
                     else:
                         self.socket.send('{}:{}'.format(self.identity, acknowledgment))
         return msg
 
     def send_to_android(self, msg):
+        '''
+        Send regular message to broker but setting channel to 'android', that
+        will make it use NotifyMyAndroid to route message toward a green robot device
+        '''
         assert isinstance(msg, dict)
         msg['time']    = dt.datetime.strftime(dt.datetime.now(), format = '%Y-%m-%dT%H:%M:%S')
         msg['type']    = 'notification'
         msg['channel'] = 'android'
         msg['appname'] = 'NeuronQuant'
-        log.debug('Client sends android notification: {}'.format(msg))
+        log.debug('Dealer sends android notification: {}'.format(msg))
         self.send(msg, format=False)
 
     def send(self, msg, format=True, **kwargs):
-        #log.debug('Client sends: %s' % msg)
+        '''
+        Sends msg through socket, taking care of missing fields in protocole
+        if format flag is set. Otherwise, autodetection
+        '''
         if format:
-            self.socket.send_json({'time': dt.datetime.strftime(dt.datetime.now(), format='%Y-%m-%dT%H:%M:%S'), 'msg': msg,
-                                   'id': self.identity, 'channel': kwargs.get('channel', 'dashboard'), 'type': kwargs.get('type', '')})
+            self.socket.send_json({'time': dt.datetime.strftime(dt.datetime.now(), format='%Y-%m-%dT%H:%M:%S'),
+                                   'msg': msg,
+                                   'id': self.identity,
+                                   'channel': kwargs.get('channel', self.recipient),
+                                   'type': kwargs.get('type', '')})
         else:
             self.socket.send_json(msg) if isinstance(msg, dict) else self.socket.send(msg)
 
 
+#TODO Unused, cleanup
 class ZMQ_Broker(threading.Thread):
     """ServerTask"""
     def __init__(self):
@@ -148,7 +212,7 @@ class ZMQ_Server(ZMQ_Base):
                 msg = self.receive()
                 try:
                     on_recv(msg, id=port)
-                    self.send({"{}:statut".format(port): 0})
+                    self.send({"{}:statut".format(port): OK_CODE})
                 except:
                     log.error('** Processing message received')
                     self.send({"{}:statut".format(port): 1})
@@ -177,6 +241,7 @@ def handle_json(msg, id):
     print(json.dumps(msg, indent=4, separators=(',', ': ')))
 
 
+#TODO Externalize tests
 def server_test():
     server = ZMQ_Server()
     server.run_forever(ports=5555, on_recv=handle_json)